Filtering and searching for bulk / non-bulk requests
to finish off this issue.
