+++ lib/Constants.js
@@ -1,5 +1,4 @@
 import { Enum } from '@/lib/ClassUtils';
-import TimeFormatters from '@/lib/time/TimeFormatters';
 
 /**
  * Authentication scopes.  Used to restrict access to certain REST API endpoints,
@@ -647,50 +646,6 @@ const FEATURE_CODES_INTERSECTION = RoadIntersectionType.enumValues
 const FEATURE_CODES_SEGMENT = RoadSegmentType.enumValues
   .map(({ featureCode }) => featureCode);
 
-const SearchKeys = {
-  Requests: {
-    ASSIGNED_TO: (q, r) => {
-      const qLower = q.toLowerCase();
-      let rLower = 'none';
-      if (r.studyRequest.assignedTo) {
-        rLower = r.studyRequest.assignedTo.text.toLowerCase();
-      }
-      return rLower.indexOf(qLower) !== -1;
-    },
-    CREATED_AT: (q, r) => {
-      const qLower = q.toLowerCase();
-      const createdAtStr = TimeFormatters.formatDefault(r.studyRequest.createdAt);
-      const rLower = createdAtStr.toLowerCase();
-      return rLower.indexOf(qLower) !== -1;
-    },
-    LOCATION: (q, r) => {
-      if (!r.location || !r.location.description) {
-        return q === '';
-      }
-      const qLower = q.toLowerCase();
-      const rLower = r.location.description.toLowerCase();
-      return rLower.indexOf(qLower) !== -1;
-    },
-    REQUESTER: (q, r) => {
-      if (r.requestedBy === null) {
-        return q === '';
-      }
-      const qLower = q.toLowerCase();
-      const rLower = r.requestedBy.toLowerCase();
-      return rLower.indexOf(qLower) !== -1;
-    },
-    STUDY_TYPE: (q, r) => {
-      const qLower = q.toLowerCase();
-      let rLower = r.studyRequest.studyType.label.toLowerCase();
-      if (rLower.indexOf(qLower) !== -1) {
-        return true;
-      }
-      rLower = r.studyRequest.studyType.name.toLowerCase();
-      return rLower.indexOf(qLower) !== -1;
-    },
-  },
-};
-
 const SignalType = {
   NORMAL: 1,
   PEDCROSS: 2,
@@ -1028,7 +983,6 @@ const Constants = {
   ReportType,
   RoadIntersectionType,
   RoadSegmentType,
-  SearchKeys,
   SignalType,
   SortDirection,
   SPEED_CLASSES,
@@ -1070,7 +1024,6 @@ export {
   ReportType,
   RoadIntersectionType,
   RoadSegmentType,
-  SearchKeys,
   SignalType,
   SortDirection,
   SPEED_CLASSES,
